changing dev environments

+++ robot/table.go
@@ -4,11 +4,11 @@ import "fmt"
 
 //Table struct has Width (Y representation) and Length (X representation)
 type Table struct {
-	Width, Length int
+	Width, Length float32
 }
 
 //NewTable is to initantiate Table object
-func NewTable(width int, length int) (Table, error) {
+func NewTable(width float32, length float32) (Table, error) {
 	if width < 0 || length < 0 {
 		return Table{}, fmt.Errorf(errFailToInitiate)
 	}
